Temporarily hide instructions for Slack
I had to turn off the self-invite system for Slack.

We received a ton of spam-ish registrations to emails that all bounced.

This also started racking up the bill for the first time for this
account, so I decided to shut it down for now.

When I have time to look into it, I'll turn it back on.